% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HB_axe.R
\name{axe_env_bayesian}
\alias{axe_env_bayesian}
\title{Adapted from butcher::axe_env for bayesian(engine='brms')}
\usage{
axe_env_bayesian(x, verbose = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Adapted from butcher::axe_env for bayesian(engine='brms')
}
