% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null.R
\name{calc_null}
\alias{calc_null}
\title{Calculate null model predictions}
\usage{
calc_null(obs.ls, resp)
}
\arguments{
\item{obs.ls}{A list of data frames containing observations for different response variables.}

\item{resp}{A character string specifying the response variable (e.g., `"alert"`, `"tl"`, `"lnN"`).}
}
\value{
A list containing two data frames: `yday.df` with null model predictions for each day of the year, and `obs.df` with the original observations and corresponding null model predictions.
}
\description{
This function calculates null model predictions for different response variables by averaging observations within a specified time window.
}
\examples{
\dontrun{
library(tidyverse)
obs.ls <- list(alert = data.frame(date = Sys.Date() - 1:10, alert = sample(c("A1", "A2"), 10, replace = TRUE)))
result <- calc_null(obs.ls, "alert")
}
}
