% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_R2}
\alias{calc_R2}
\title{Calculate pseudo-R2s}
\usage{
calc_R2(dat.df, type = "mf", ...)
}
\arguments{
\item{type}{}
}
\description{
Currently supports McFadden and the Veall-Zimmermann correction of the Aldrich-Nelson pseudo-R2
}
