% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ydayAvg.R
\name{calc_ydayAvg}
\alias{calc_ydayAvg}
\title{Calculate average for each day of year}
\usage{
calc_ydayAvg(data_df, out_f, ...)
}
\arguments{
\item{data_df}{Dataframe}

\item{out_f}{Output file}

\item{...}{Columns to group by}
}
\value{
Written object to out_f only
}
\description{
Calculate average for each day of year
}
