% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_thresholds}
\alias{compute_thresholds}
\title{Compute classification metrics across probability thresholds}
\usage{
compute_thresholds(
  L.df,
  prMin = 0,
  prMax = 1,
  prSteps = 0.1,
  byPrevAlert = F,
  cores = 1
)
}
\arguments{
\item{prSteps}{}
}
\description{
Compute classification metrics across probability thresholds
}
