% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFoldsByYear.R
\name{createFoldsByYear}
\alias{createFoldsByYear}
\title{Identify indexes for ML training}
\usage{
createFoldsByYear(data.df)
}
\arguments{
\item{data.df}{}
}
\description{
Identify indexes for ML training
}
