% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dirrf}
\alias{dirrf}
\title{Shortcut for dir(..., full.names=T, recursive=T)}
\usage{
dirrf(...)
}
\arguments{
\item{...}{Arguments to `dir()`. Usually path and possibly pattern}
}
\value{
A character vector containing the names of the files in the specified
directories (empty if there were no files). If a path does not exist or is
not a directory or is unreadable it is skipped.
}
\description{
Shortcut for dir(..., full.names=T, recursive=T)
}
