% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_sites.R
\name{download_site_info}
\alias{download_site_info}
\title{This function downloads site data from the HABReports database.}
\usage{
download_site_info(url_sites, dateStart)
}
\arguments{
\item{url_sites}{A character string specifying the URL of the site table in the HABReports database.}

\item{dateStart}{A Date object specifying the start date; sites operating only before this date will be ignored.}
}
\value{
A data frame with columns: sin, site, area, farm_species, east, north, fromdate, and todate.
}
\description{
This function downloads site data from the HABReports database.
}
\examples{
# Example usage:
# url_sites <- "http://example.com/sites"
# dateStart <- as.Date("2023-01-01")
# download_site_info(url_sites, dateStart)
}
