% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_env.R
\name{extract_env_pts}
\alias{extract_env_pts}
\title{Extract CMEMS or WRF data to site point locations}
\usage{
extract_env_pts(site.df, env_vars, env.df, id_env, site.v)
}
\arguments{
\item{site.v}{}
}
\description{
Extract CMEMS or WRF data to site point locations
}
