% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_env.R
\name{find_buffer_intersect_ids}
\alias{find_buffer_intersect_ids}
\title{Find ids of environmental layer points within each site buffer}
\usage{
find_buffer_intersect_ids(site.sf, env.sf, id_env)
}
\arguments{
\item{site.sf}{An `sf` object with site quadrants.}

\item{env.sf}{An `sf` object with environmental features.}

\item{id_env}{A character string specifying the ID column in `env.sf`.}
}
\value{
A data frame with the intersecting environmental feature IDs added.
}
\description{
This function finds the IDs of environmental layer points that fall within each site buffer.
}
