% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_env.R
\name{find_nearest_feature_id}
\alias{find_nearest_feature_id}
\title{Find id of environmental layer points nearest to each site point location}
\usage{
find_nearest_feature_id(site.df, env.sf, id_env)
}
\arguments{
\item{id_env}{}
}
\description{
Find id of environmental layer points nearest to each site point location
}
