% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_candidate.R
\name{fit_candidate}
\alias{fit_candidate}
\title{Wrapper to fit a model}
\usage{
fit_candidate(mod, resp, form.ls, d.ls, opts, tunes, out.dir, y, suffix = NULL)
}
\arguments{
\item{suffix}{}
}
\description{
Wrapper to fit a model
}
