% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{get_lags}
\alias{get_lags}
\title{Lag multiple variables at once}
\usage{
get_lags(data, ..., n = 2)
}
\arguments{
\item{data}{Dataframe}

\item{...}{Unquoted variable names to lag}

\item{n}{Number of lags}
}
\description{
From https://stackoverflow.com/questions/55814028/multiple-lags-with-dplyr
}
