% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null.R
\name{get_rows_by_yday}
\alias{get_rows_by_yday}
\title{Identify rows in y within thresh days of x}
\usage{
get_rows_by_yday(x, y, thresh)
}
\arguments{
\item{thresh}{}
}
\description{
Identify rows in y within thresh days of x
}
