% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_geography.R
\name{get_shortestPaths}
\alias{get_shortestPaths}
\title{Find pairwise shortest in-ocean paths between sites}
\usage{
get_shortestPaths(ocean.path, site.df, site_savePath = NULL)
}
\arguments{
\item{recalc_transMx}{}
}
\description{
Find pairwise shortest in-ocean paths between sites
}
