% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tl_info.R
\name{get_tl_info}
\alias{get_tl_info}
\title{Get Traffic light threshold level information}
\usage{
get_tl_info(target_i)
}
\arguments{
\item{target_i}{A tibble containing the target parameters to be joined with the threshold level information.}
}
\value{
A tibble with the combined and processed threshold level information.
}
\description{
This function retrieves and processes threshold level information for a given set of target parameters.
}
\examples{
target_i <- tibble(hab_parameter = c("param1", "param2"), abbr = c("P1", "P2"))
get_tl_info(target_i)
}
