% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{get_trafficLights}
\alias{get_trafficLights}
\title{Identify traffic light and alert status based on density/concentration}
\usage{
get_trafficLights(y.df, N, tl_i)
}
\arguments{
\item{tl_i}{}
}
\description{
Identify traffic light and alert status based on density/concentration
}
