% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_SMOTE.R
\name{knearest}
\alias{knearest}
\title{Find k-Nearest Neighbors}
\usage{
knearest(D, P, n_clust)
}
\arguments{
\item{D}{A data frame or matrix representing the dataset.}

\item{P}{A data frame or matrix representing the points for which to find the nearest neighbors.}

\item{n_clust}{The number of nearest neighbors to find.}
}
\value{
A matrix where each row contains the indices of the k-nearest neighbors for each point in P.
}
\description{
This function finds the k-nearest neighbors for each point in the dataset using the FNN package.
}
\examples{
\dontrun{
D <- matrix(runif(100), ncol = 2)
P <- matrix(runif(20), ncol = 2)
n_clust <- 3
neighbors <- knearest(D, P, n_clust)
}
}
