% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_recipe.R
\name{load_dataset_y}
\alias{load_dataset_y}
\title{Load dataframe for specified target}
\usage{
load_dataset_y(data.dir, y.i, col_metadata, col_resp, all_covs)
}
\arguments{
\item{all_covs}{}
}
\description{
Load dataframe for specified target
}
