% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{load_datasets}
\alias{load_datasets}
\title{Load datasets for compilation}
\usage{
load_datasets(sub.dir, target, ydayAvg.dir)
}
\arguments{
\item{sub.dir}{A character string specifying the subdirectory where the data files are located.}

\item{target}{A character string specifying the target dataset (e.g., "tox").}

\item{ydayAvg.dir}{A character string specifying the directory where the year-day average data files are located.}
}
\value{
A list of data frames containing the compiled datasets.
}
\description{
This function loads and compiles various datasets for a specified target, including site data, observations, CMEMS data, WRF data, and additional environmental data.
}
\examples{
\dontrun{
library(tidyverse)
sub.dir <- "subdirectory"
target <- "tox"
ydayAvg.dir <- "ydayAvg_directory"
datasets <- load_datasets(sub.dir, target, ydayAvg.dir)
}
}
