% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_recipe.R
\name{make_all_covs}
\alias{make_all_covs}
\title{Make all covariates}
\usage{
make_all_covs(col_cmems, col_wrf, y_i)
}
\arguments{
\item{y_i}{}
}
\value{
list with `spacetime`,`main`, `interact`, and `hab`, each containing a vector of covariate names
}
\description{
Make all covariates
}
