% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast_data.R
\name{make_forecast_data}
\alias{make_forecast_data}
\title{Make forecast data}
\usage{
make_forecast_data(y_i, covSet, data.dir, responses = c(alert = "alert"))
}
\arguments{
\item{responses}{}
}
\description{
This function loads compiled datasets and applies the appropriate recipe to
prep a new dataset for a specific covSet.
}
