% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HB_setup.R
\name{make_HB_formula}
\alias{make_HB_formula}
\title{Create formulas for Hierarchical Bayesian models}
\usage{
make_HB_formula(resp, covs, splinesInt = "both")
}
\arguments{
\item{splinesInt}{}
}
\description{
Create formulas for Hierarchical Bayesian models
}
