% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{merge_pred_dfs}
\alias{merge_pred_dfs}
\title{Merge summarised predictions across all models into single dataframes}
\usage{
merge_pred_dfs(files, CV = NULL)
}
\arguments{
\item{files}{A vector of full file paths to the prediction files.}

\item{CV}{A character string specifying the cross-validation type. Options are "HB", "ML", or NULL. Default is NULL.}
}
\value{
A list of data frames with the merged predictions.
}
\description{
This function merges summarised predictions across all models into single dataframes.
}
\examples{
\dontrun{
files <- c("path/to/predictions1.rds", "path/to/predictions2.rds")
merged_preds <- merge_pred_dfs(files)
}
}
