% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_SMOTE.R
\name{n_dup_max}
\alias{n_dup_max}
\title{Calculate Maximum Number of Duplicates}
\usage{
n_dup_max(size_input, size_P, size_N, dup_size = 0)
}
\arguments{
\item{size_input}{The total size of the input dataset.}

\item{size_P}{The size of the positive (minority) class.}

\item{size_N}{The size of the negative (majority) class.}

\item{dup_size}{The desired duplication size. Default is 0.}
}
\value{
The maximum number of duplicates to generate.
}
\description{
This function calculates the maximum number of duplicates to generate based on the input sizes and duplication size.
}
\examples{
\dontrun{
size_input <- 1000
size_P <- 50
size_N <- 950
dup_size <- 0
max_dup <- n_dup_max(size_input, size_P, size_N, dup_size)
}
}
