% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_geography.R
\name{points2nearestcell}
\alias{points2nearestcell}
\title{Shift out-of-bounds pointsto nearest cell}
\usage{
points2nearestcell(
  locs = NULL,
  ras = NULL,
  layer = 1,
  move = T,
  distance = NULL,
  showchanges = T
)
}
\arguments{
\item{showchanges}{}
}
\description{
Shift out-of-bounds pointsto nearest cell
}
