% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_SMOTE.R
\name{rebalance_smote}
\alias{rebalance_smote}
\title{Generate Synthetic Samples Using Modified SMOTE}
\usage{
rebalance_smote(og_df, K = 5, dup_size = 0)
}
\arguments{
\item{og_df}{A data frame containing the original dataset.}

\item{K}{The number of nearest neighbors to use for generating synthetic samples.}

\item{dup_size}{The desired size of the synthetic dataset.}
}
\value{
A data frame containing the original and synthetic data.
}
\description{
This function generates synthetic samples for the minority class using a modified version of SMOTE
from the smotefamily package. It then combines the original and synthetic data.
}
\examples{
\dontrun{
og_df <- data.frame(
  siteid = rep(1:10, each = 10),
  lon = runif(100, -180, 180),
  lat = runif(100, -90, 90),
  obsid = 1:100,
  y = rnorm(100),
  date = Sys.Date() - 1:100,
  alert = sample(c("A1", "A2"), 100, replace = TRUE),
  alert1 = sample(1:2, 100, replace = TRUE),
  alert2 = sample(1:2, 100, replace = TRUE),
  tl = sample(1:5, 100, replace = TRUE)
)
K <- 5
dup_size <- 100
result <- generate_synthetic_samples(og_df, K, dup_size)
}
}
