% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{schoenr_alt}
\alias{schoenr_alt}
\title{Schoener's D calculation that deals with NAs better}
\usage{
schoenr_alt(d1, d2, a = NULL, b = NULL)
}
\arguments{
\item{b}{}
}
\description{
Schoener's D calculation that deals with NAs better
}
