% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{split_to_NSEW}
\alias{split_to_NSEW}
\title{Split circular buffer into NSEW quadrants}
\usage{
split_to_NSEW(sf, radius = 110000)
}
\arguments{
\item{radius}{}
}
\description{
Split circular buffer into NSEW quadrants
}
