% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WRF.R
\name{subset_WRF}
\alias{subset_WRF}
\title{Filter WRF data to include only preserved points in each domain}
\usage{
subset_WRF(domain, wrf.out, v2_start = NULL, refreshStart = NULL)
}
\arguments{
\item{v2_start}{}
}
\description{
Filter WRF data to include only preserved points in each domain
}
