% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{summarise_ML_preds}
\alias{summarise_ML_preds}
\title{Calculate mean predictions for ML models}
\usage{
summarise_ML_preds(preds, resp, y_i.i)
}
\arguments{
\item{preds}{A data frame containing the predictions.}

\item{resp}{A character string specifying the response variable.}

\item{y_i.i}{A data frame with information about the variables of interest.}
}
\value{
A data frame with the mean predictions.
}
\description{
This function calculates mean predictions for machine learning models.
}
\examples{
\dontrun{
preds <- tibble(.pred_A1 = runif(10))
resp <- "alert"
y_i.i <- data.frame(tl_thresh = "TL2", N_thresh = 1)
mean_preds <- summarise_ML_preds(preds, resp, y_i.i)
}
}
