% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{summarise_predictions}
\alias{summarise_predictions}
\title{Summarise and aggregate predictions for a set of models}
\usage{
summarise_predictions(d.y, dPCA.y, resp, fit.dir, y_i.i, suffix = NULL)
}
\arguments{
\item{suffix}{}
}
\description{
Summarise and aggregate predictions for a set of models
}
