% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HB_axe.R
\name{axe_env_bayesian}
\alias{axe_env_bayesian}
\title{Adapted from butcher::axe_env for bayesian(engine='brms')}
\usage{
axe_env_bayesian(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A fitted Bayesian model object.}

\item{verbose}{A logical value indicating whether to print detailed information during the process. Default is FALSE.}

\item{...}{Additional arguments (currently not used).}
}
\value{
The modified Bayesian model object with reduced environments.
}
\description{
This function adapts the `butcher::axe_env()` function for Bayesian models fitted with the `brms` engine, removing unnecessary environments to reduce the size of the model object.
}
\examples{
\dontrun{
library(brms)
fit <- brm(mpg ~ wt + cyl, data = mtcars)
fit_reduced <- axe_env_bayesian(fit)
}
}
