% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{calc_LL_wts}
\alias{calc_LL_wts}
\title{Calculate model weights based on mean log loss}
\usage{
calc_LL_wts(cv.df, resp, wt.penalty = 2)
}
\arguments{
\item{wt.penalty}{}
}
\description{
Calculate model weights based on mean log loss
}
