% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{calc_y_features}
\alias{calc_y_features}
\title{Calculate local and regional autoregressive terms for HAB and toxin observations}
\usage{
calc_y_features(
  yRaw.df,
  y_i,
  tl_i,
  dist.df = NULL,
  forecastStart = "3000-01-01"
)
}
\arguments{
\item{site.100km}{}
}
\description{
Calculate local and regional autoregressive terms for HAB and toxin observations
}
