% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFoldsByYear.R
\name{createFoldsByYear}
\alias{createFoldsByYear}
\title{Identify indexes for ML training}
\usage{
createFoldsByYear(data.df)
}
\arguments{
\item{data.df}{Dataframe}
}
\value{
List of 2 lists where `i.in` gives the row indexes to include in
each fold, and `i.out` gives the indexes to exclude in each fold.
}
\description{
Identify indexes for ML training
}
