% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{detrend_loess}
\alias{detrend_loess}
\title{Detrend observations using a loess smoother}
\usage{
detrend_loess(x, y, span = 0.75, robust = TRUE)
}
\arguments{
\item{robust}{}
}
\description{
Detrend observations using a loess smoother
}
