% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_env.R
\name{extract_env_buffers}
\alias{extract_env_buffers}
\title{Extract CMEMS or WRF data within site buffer quadrants}
\usage{
extract_env_buffers(site.buffer, vars, env.df, id_env)
}
\arguments{
\item{id_env}{}
}
\description{
Extract CMEMS or WRF data within site buffer quadrants
}
