% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_recipe.R
\name{filter_corr_covs}
\alias{filter_corr_covs}
\title{Filter list of covariates following recipe thinning}
\usage{
filter_corr_covs(all_covs, data.y, covsExclude = "NA")
}
\arguments{
\item{covsExclude}{}
}
\description{
Filter list of covariates following recipe thinning
}
