% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{find_AUCPR_min}
\alias{find_AUCPR_min}
\title{Find minimum possible PR-AUC}
\usage{
find_AUCPR_min(df, ...)
}
\arguments{
\item{df}{Dataframe}

\item{...}{Unquoted grouping variables}
}
\description{
Following Boyd et al 2012
}
