% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_env.R
\name{find_buffer_intersect_ids}
\alias{find_buffer_intersect_ids}
\title{Find ids of environmental layer points within each site buffer}
\usage{
find_buffer_intersect_ids(site.sf, env.sf, id_env)
}
\arguments{
\item{id_env}{}
}
\description{
Find ids of environmental layer points within each site buffer
}
