% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_covSets.R
\name{fit_covSet}
\alias{fit_covSet}
\title{Title}
\usage{
fit_covSet(
  y_i,
  run_type = "0_init",
  covSet,
  mod,
  test_prop = 0.75,
  nTuneVal = 2,
  prior_strength = 1,
  ncores = 4,
  responses = c(alert = "alert")
)
}
\arguments{
\item{responses}{}
}
\description{
Title
}
