% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ensemble.R
\name{fit_ensemble}
\alias{fit_ensemble}
\title{Calculate ensemble model predictions}
\usage{
fit_ensemble(
  out.ls,
  wt.ls,
  resp,
  y_i.i,
  method = "wtmean",
  out.path = NULL,
  opt = NULL
)
}
\arguments{
\item{y_i.i}{}
}
\description{
Calculate ensemble model predictions
}
