% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CMEMS.R
\name{get_CMEMS}
\alias{get_CMEMS}
\title{Download CMEMS layers}
\usage{
get_CMEMS(
  userid,
  pw,
  i.df,
  bbox,
  nDays_buffer,
  dateRng,
  out.dir,
  toolbox = TRUE
)
}
\arguments{
\item{out.dir}{}
}
\description{
Download CMEMS layers
}
