% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_recipe.R
\name{get_excluded_cov_regex}
\alias{get_excluded_cov_regex}
\title{Get regex for excluded covariates given a covariate set}
\usage{
get_excluded_cov_regex(covSet)
}
\arguments{
\item{covSet}{}
}
\description{
Get regex for excluded covariates given a covariate set
}
