% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_geography.R
\name{get_fetch}
\alias{get_fetch}
\title{Extract fetch for each site point location}
\usage{
get_fetch(site.df, fetch.path)
}
\arguments{
\item{fetch.path}{}
}
\description{
Extract fetch for each site point location
}
