% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Calculate J, F-1, precision, and recall for a dataframe}
\usage{
get_metrics(dat)
}
\arguments{
\item{dat}{}
}
\description{
This is a function to pass to future_map to avoid capturing large objects
in the environment which makes it extraordinarily slow
}
