% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_geography.R
\name{get_openBearing}
\alias{get_openBearing}
\title{Identify most-open bearing for each site point location}
\usage{
get_openBearing(site.df, coast.path, buffer = 10000, nDir = 120)
}
\arguments{
\item{nDir}{}
}
\description{
Identify most-open bearing for each site point location
}
