% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Generate predictions for a model}
\usage{
get_predictions(fit, mod, resp, d.df, y_i.i)
}
\arguments{
\item{y_i.i}{}
}
\description{
Generate predictions for a model
}
