% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HB_run_CV.R
\name{HB_run_CV}
\alias{HB_run_CV}
\title{Run cross-validation for Bayesian models}
\usage{
HB_run_CV(
  mod,
  folds,
  cv.dir,
  y,
  y_i.i,
  r,
  form.ls,
  HB.i,
  priors,
  PCA = F,
  reverse = F
)
}
\arguments{
\item{priors}{}
}
\description{
Run cross-validation for Bayesian models
}
