% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{load_datasets}
\alias{load_datasets}
\title{Load datasets for compilation}
\usage{
load_datasets(sub.dir, target, ydayAvg.dir)
}
\arguments{
\item{target}{}
}
\description{
Load datasets for compilation
}
