% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HB_setup.R
\name{make_HB_priors}
\alias{make_HB_priors}
\title{Create priors for each Hierarchical Bayesian model}
\usage{
make_HB_priors(prior_i, mod, resp, covs, PCA = F)
}
\arguments{
\item{covs}{}
}
\description{
Create priors for each Hierarchical Bayesian model
}
