% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{merge_pred_dfs}
\alias{merge_pred_dfs}
\title{Merge summarised predictions across all models into single dataframes}
\usage{
merge_pred_dfs(files, CV = NULL)
}
\arguments{
\item{files}{vector of full file path}
}
\description{
Merge summarised predictions across all models into single dataframes
}
