% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WRF.R
\name{nest_WRF_domains}
\alias{nest_WRF_domains}
\title{Identify preserved points within nested WRF domains}
\usage{
nest_WRF_domains(nc.ls)
}
\arguments{
\item{nc.ls}{}
}
\description{
Identify preserved points within nested WRF domains
}
