% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_geography.R
\name{point_in_cell}
\alias{point_in_cell}
\title{Identify whether point locations fall within valid raster cells}
\usage{
point_in_cell(locs, ras, layer = 1)
}
\arguments{
\item{layer}{}
}
\description{
Identify whether point locations fall within valid raster cells
}
