% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_recipe.R
\name{prep_recipe}
\alias{prep_recipe}
\title{Create recipe and prepare using training data}
\usage{
prep_recipe(train.df, response, covsExclude = "NA", dimReduce = FALSE)
}
\arguments{
\item{dimReduce}{}
}
\description{
Create recipe and prepare using training data
}
