% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_sites.R
\name{read_and_clean_fish}
\alias{read_and_clean_fish}
\title{Title}
\usage{
read_and_clean_fish(url_mowi, url_ssf, fish_i, sites, dateStart = "2016-01-01")
}
\arguments{
\item{dateStart}{}
}
\description{
Title
}
