% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_sites.R
\name{read_and_clean_sites}
\alias{read_and_clean_sites}
\title{Read and clean sites}
\usage{
read_and_clean_sites(url_sites, dateStart)
}
\arguments{
\item{url_sites}{URL of site table in HABReports database}

\item{dateStart}{Start date; ignores sites only operating before this}
}
\value{
data.frame with sin, east, north, and date
}
\description{
Read and clean sites
}
