% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{split_to_NSEW}
\alias{split_to_NSEW}
\title{Split circular buffer into NSEW quadrants}
\usage{
split_to_NSEW(sf, radius = 110000)
}
\arguments{
\item{sf}{An `sf` object representing the circular buffer.}

\item{radius}{Numeric value specifying the radius of the buffer in meters. Default is 110,000 meters.}
}
\value{
An `sf` object with the circular buffer split into NSEW quadrants.
}
\description{
This function splits a circular buffer into four quadrants: North, South, East, and West.
}
\examples{
\dontrun{
library(sf)
# Create a sample sf object
sf <- st_as_sf(data.frame(siteid = 1, geometry = st_sfc(st_point(c(0, 0)))), crs = 27700)
# Split the buffer into quadrants
quadrants <- split_to_NSEW(sf, radius = 110e3)
}
}
