% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{summarise_ML_preds}
\alias{summarise_ML_preds}
\title{Calculate mean predictions for ML models}
\usage{
summarise_ML_preds(preds, resp, y_i.i)
}
\arguments{
\item{y_i.i}{}
}
\description{
Calculate mean predictions for ML models
}
