% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{summarise_post_preds}
\alias{summarise_post_preds}
\title{Calculate mean predictions for HB models}
\usage{
summarise_post_preds(post, resp, y_i.i)
}
\arguments{
\item{y_i.i}{}
}
\description{
Calculate mean predictions for HB models
}
